\name{stat-package}
\alias{stat-package}
\alias{stat}
\docType{package}
\title{
Calculate mean and variance in short time.
}
\description{
Within shor time calculate mean and variance.
\details{
\tabular{ll}{
Package: \tab stat\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-08\cr
License: \tab GPL-2\cr
}
}
\author{
Md. Mushfiqur Rahman

Maintainer: Md. Mushfiqur Rahamn <mushfiq_194@yahoo.com>
}
\references{
Md Nurul haque Mollah
\keyword{ package }

\examples{

}
